saliency heatmaps

+++ icairdweakly/create_saliency_heatmaps.py
@@ -464,7 +464,7 @@ if __name__ == '__main__':
                 print(an_scale_x, an_scale_y)
                 an_x, an_x1, an_y, an_y1 = min_x//an_scale_x, max_x//an_scale_x, min_y//an_scale_y, max_y//an_scale_y
                 print(an_x, an_x1, an_y, an_y1)
-                an_x_dim, an_y_dim = an_scale_x*(an_x1 - an_x), an_scale_y*(an_y1 - an_y)
+                an_x_dim, an_y_dim = int(an_scale_x*(an_x1 - an_x)), int(an_scale_y*(an_y1 - an_y))
                 scaled_an = F.interpolate(img[:,:,an_x:an_x1, an_y:an_y1], (an_x_dim, an_y_dim))[0]
                 print(scaled_an.shape)
                 print(full_img.shape)
